#44 [REFACTOR] queryClient를 lib에서 providers로 이동
